Done Order Function

+++ KoiCareSystematHome/KoiCareSystem.Common/DTOs/Request/RequestCreateANewProductDto.cs
@@ -9,11 +9,11 @@ namespace KoiCareSystem.Common.DTOs.Request
 {
     public class RequestCreateANewProductDto
     {
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductType { get; set; }
         public decimal? Price { get; set; }
-        public long? StockQuantity { get; set; }
-        public long CategoryId { get; set; }
+        public int? StockQuantity { get; set; }
+        public int CategoryId { get; set; }
     }
 }
