+++ KoiCareSystematHome/KoiCareSystem.Common/DTOs/Request/RequestItemToOrderDto.cs
@@ -8,13 +8,13 @@ namespace KoiCareSystem.Common.DTOs.Request
 {
     public class RequestItemToOrderDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
-        public long OrderId { get; set; }
+        public int OrderId { get; set; }
 
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
 
-        public long Quantity { get; set; }
+        public int Quantity { get; set; }
 
     }
 }
