+++ KoiCareSystematHome/KoiCareSystem.RazorWebApp/Pages/Admin/Orders/Delete.cshtml.cs
@@ -23,14 +23,14 @@ namespace KoiCareSystem.RazorWebApp.Pages.Admin.Orders
         [BindProperty]
         public Order Order { get; set; } = default!;
 
-        public async Task<IActionResult> OnGetAsync(long? id)
+        public async Task<IActionResult> OnGetAsync(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            var order = await _orderService.GetOrderByOrderId((long)id);
+            var order = await _orderService.GetOrderByOrderId((int)id);
 
             if (order == null)
             {
@@ -43,17 +43,17 @@ namespace KoiCareSystem.RazorWebApp.Pages.Admin.Orders
             return Page();
         }
 
-        public async Task<IActionResult> OnPostAsync(long? id)
+        public async Task<IActionResult> OnPostAsync(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            var order = await _orderService.GetOrderByOrderId((long)id);
+            var order = await _orderService.GetOrderByOrderId((int)id);
             if (order != null)
             {
-                await _orderService.DeleteOrderByOrderId((long)id);
+                await _orderService.DeleteOrderByOrderId((int)id);
             }
 
             return RedirectToPage("./Index");
