+++ KoiCareSystematHome/KoiCareSystem.RazorWebApp/Pages/Admin/Orders/Details.cshtml.cs
@@ -22,14 +22,14 @@ namespace KoiCareSystem.RazorWebApp.Pages.Admin.Orders
 
         public Order Order { get; set; } = default!;
 
-        public async Task<IActionResult> OnGetAsync(long? id)
+        public async Task<IActionResult> OnGetAsync(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            var order = await _orderService.GetOrderByOrderId((long)id);
+            var order = await _orderService.GetOrderByOrderId((int)id);
             if (order == null)
             {
                 return NotFound();
