+++ KoiCareSystematHome/KoiCareSystem.RazorWebApp/Pages/Admin/Products/Edit.cshtml.cs
@@ -30,14 +30,14 @@ namespace KoiCareSystem.RazorWebApp.Pages.Admin.Products
         [BindProperty]
         public RequestCreateANewProductDto RequestCreateANewProductDto { get; set; } = default!;
         public Product Product { get; set; } = default!;
-        public async Task<IActionResult> OnGetAsync(long? id)
+        public async Task<IActionResult> OnGetAsync(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            var product =  await _productService.GetProductById((long)id);
+            var product =  await _productService.GetProductById((int)id);
             if (product == null)
             {
                 return NotFound();
